feat: 구인글 등록 테스트코드
